refactor: 124-rename folder to kebab-case and move it to components
